+++ Assets/Scripts/Arena/BattleSystem/Weapons/Weapons/Gun/Bullet.cs
@@ -19,7 +19,7 @@ public class Bullet : Projectile
         }
     }
 
-    public void SetUp(Vector3 direction, Vector3 position, int attackDamage)
+    public void SetUp(Vector3 direction, Vector3 position, float attackDamage)
     {
         SetDirection(direction);
         SetPosition(position);
@@ -36,7 +36,7 @@ public class Bullet : Projectile
         transform.position = position;
     }
 
-    private void SetAttackDamage(int attackDamage)
+    private void SetAttackDamage(float attackDamage)
     {
         this.attackDamage = attackDamage;
     }
